frontend implied

+++ database/migrations/2023_08_09_162326_create_cart_items.php
@@ -11,20 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('cart_items', function (Blueprint $table) {
-            $table->id();
-            // cart_id
-            // product_id
-    
-             $table->unsignedBigInteger('cart_id');
-            $table->unsignedBigInteger('product_id');
-            $table->integer('quantity');
-            $table->timestamps();
-            $table->dropIndex('cart_id');
-
-            $table->foreign('cart_id')->references('id')->on('carts')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products');
-        });
+        if (!Schema::hasTable('cart_items')) {
+            Schema::create('cart_items', function (Blueprint $table) {
+                $table->id();
+                 $table->unsignedBigInteger('cart_id');
+                $table->unsignedBigInteger('product_id');
+                $table->integer('quantity');
+                $table->timestamps();
+                $table->foreign('cart_id')->references('id')->on('carts')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products');
+            });
+        }
+        
     }
 
     /**
